webview.js made to hold webview related logic
