+++ app/index.html
@@ -14,7 +14,7 @@
 		</script>
 		<script type='text/javascript' src='js/prereq.js'></script>
 		<script type='text/javascript' src='js/UI.js'></script>
-		<script type='text/javascript' src='js/util.js'></script>
+		<script type='text/javascript' src='js/webview.js'></script>
 		<script type='text/javascript' src='js/plugins.js'></script>
 		<script>
 			$(window).load(function() {
